Added get_sup_groups().
This is a native Python function. Providing bindings for FUSE's native
fuse_req_getgroups() did not seem like a good idea, because it requires
to keep a reference to the FUSE request object in the Python RequestContext
object. Since the latter may outlive the former, we would need to introduce
locking to prevent the request object to be freed while a Python function
is using it.

Implementing the function in Python and passing a pid instead of a request
object circumvents these problems nicely. The solution is still linux specific
(because it parses /proc/pid/status), but it seems that neither OSXFUSE
nor FUSE4BSD provide implementations of fuse_req_getgroups() either, so
we're not loosing anything.

Fixes issue #3.